+++ app/Http/Controllers/GroupPermissionController.php
@@ -14,6 +14,9 @@ class GroupPermissionController extends Controller
      */
     public function index()
     {
+        if (!$this->CheckPermission("view_group_permissions", 1)) {
+            return $this->sendError($error = 'Unauthorized', $code = 403);
+        }
         $groupPermissions = GroupPermission::with('group')->with('group.users')->paginate(20);
 
         return $this->sendResponse(GroupPermissionResource::collection($groupPermissions)
@@ -25,6 +28,9 @@ class GroupPermissionController extends Controller
      */
     public function store(Request $request)
     {
+        if (!$this->CheckPermission("add_group_permission", 1)) {
+            return $this->sendError($error = 'Unauthorized', $code = 403);
+        }
         $input = $request->all();
 
         $validator = Validator::make($input, [
@@ -47,6 +53,9 @@ class GroupPermissionController extends Controller
      */
     public function show(string $id)
     {
+        if (!$this->CheckPermission("view_group_permission", 1)) {
+            return $this->sendError($error = 'Unauthorized', $code = 403);
+        }
         $groupPermission = GroupPermission::with('group')->with('group.users')->find($id);
 
         if (is_null($groupPermission)) {
@@ -62,6 +71,9 @@ class GroupPermissionController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (!$this->CheckPermission("update_group_permissions", 1)) {
+            return $this->sendError($error = 'Unauthorized', $code = 403);
+        }
         $validator = Validator::make($request->all(), [
             'group_id' => 'required',
             'folder_id' => 'required'
@@ -88,6 +100,9 @@ class GroupPermissionController extends Controller
      */
     public function destroy(string $id)
     {
+        if (!$this->CheckPermission("delete_group_permissions", 1)) {
+            return $this->sendError($error = 'Unauthorized', $code = 403);
+        }
         GroupPermission::find($id)->delete();
 
         return $this->sendResponse([], 'Group Permission deleted successfully.');
